Fix a minor bug in people/person.c
When asked to enter the sex of a person, if the user entered a non
digit value, the function would print two error message. One because
the input wasn't a digit, and one because the numerical value of the
input was not 0 or 1.

This is now fixed, so that only one error message prints.